pyproject.toml
[project]
name = "terminal2f"
version = "0.1.0"
description = "terminal2F, gustofied reborn"
readme = "README.md"
authors = [
    { name = "gustofied", email = "gustofied@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "mistral-common>=1.8.8",
    "mistralai>=1.10.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rerun-sdk>=0.28.1",
    "tqdm>=4.67.1",
    "typer>=0.21.1",
]

[project.scripts]
t2f = "terminal2f.main:main"


[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"
